scripts/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile scripts/requirements.in -o scripts/requirements.txt
annotated-types==0.7.0
    # via pydantic
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
dirtyjson==1.0.8
    # via -r scripts/requirements.in
idna==3.7
    # via requests
jinja2==3.1.4
    # via -r scripts/requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pillow==10.3.0
    # via -r scripts/requirements.in
pydantic==2.7.1
    # via -r scripts/requirements.in
pydantic-core==2.18.2
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via -r scripts/requirements.in
pyyaml==6.0.1
    # via -r scripts/requirements.in
requests==2.32.2
    # via -r scripts/requirements.in
rich==13.7.1
    # via
    #   -r scripts/requirements.in
    #   typer
shellingham==1.5.4
    # via typer
typer==0.12.3
    # via -r scripts/requirements.in
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
urllib3==2.2.1
    # via requests
